backend/requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Telegram Bot
python-telegram-bot==21.7
httpx>=0.27.0,<0.28.0
beautifulsoup4==4.12.3
lxml==5.2.2
feedparser==6.0.11

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0

# AI/LLM
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
openai==1.54.4

# PyTorch CPU-only (install before sentence-transformers to avoid CUDA deps)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# Memory & Vector Store
chromadb==0.5.20
sentence-transformers==3.3.1
numpy==1.26.4
pandas==2.2.2

# Scheduling
apscheduler==3.10.4

# WebSocket
websockets==14.1